v2.2.1
Estimated Position working but some of the values are a bit flipped

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/PSRobotLibs/lib/hardware/PSMotor.java
@@ -107,7 +107,7 @@ public class PSMotor {
         return wheelSize * Math.PI * rotations;
     }
 
-    public double getEncoderPosition() {
+    public int getEncoderPosition() {
         return motorObject.getCurrentPosition();
     }
